CHROMIUMOS: bgscan: Add centralized signal_monitor
The signal strength number tells only half of the picture with
respect to signal quality.  The other half is the prevailing
noise on the channel.  The two together indicate how easily the
interface is able to pull data out of the background noise.

This change takes noise into account, by caclulating a CQM
threshold based both on the RSSI and noise.  This noise value is
available in the CQM events as well as polled data.  Since most
background scan algorithms will probably want to do this the same
way, I have added this as routines in bgscan.c that can be
utilized by individual background scan mechanisms.

Additionally, I've modified bgscan_simple to use this mechanism,
and scheduled noise floor updates to happen at the same rate as
background scans (the noisefloor should be a slowly changing
parameter).

BUG=chromium:190663
TEST=Manual -- observe nl80211 messages within supplicant indicating
     actual CQM thresholds, as well as timing between these and bgscan
     messages indicating updates.

Change-Id: I901a88d5b8a34b81db43a8cfc210722fb41a6ad3
Reviewed-on: https://chromium-review.googlesource.com/323774
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>