CHROMIUMOS: wpa_supplicant: improve roaming when noise floor available
When the noise floor is available we can do a better job deciding when
to roam by using SNR instead of the signal level.  In addition roam only
if the current AP's snr is below a threshold.  For situations where the
noise floor is not available we fall back to the previous algorithm.
New config parameters roam_threshold and roam_min hold the values used
by the roaming algorithm. These can also be set through the control
interface.

BUG=chromium:201970
TEST=WiFiRoaming suite + walk-around tests + ATEN roaming tests

Original-reviewed-on: https://gerrit.chromium.org/gerrit/11982
Original-reviewed-by: Sam Leffler <sleffler@chromium.org>
Original-tested-by: Sam Leffler <sleffler@chromium.org>
Change-Id: I82dbaa7b84e138a19f11004c066db913fc33ba7f
Reviewed-on: https://chromium-review.googlesource.com/324040
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>