CHROMIUMOS: wpa_supplicant: Don't leak scan frequencies
Use a single buffer for storing the list of frequencies to use
for background scan (in the order they should be scanned).
Instead of allocating it once for each scan (then leaking it,
as the previous implementation did), create the buffer at
the start of the background scan method.

BUG=chromium:212512
TEST=valgrind

Original-reviewed-on: https://gerrit.chromium.org/gerrit/24829
Original-reviewed-by: Gary Morain <gmorain@chromium.org>
Original-tested-by: Paul Stewart <pstew@chromium.org>
Change-Id: I69d4a792321cb561f08d638f408f848545ca1274
Reviewed-on: https://chromium-review.googlesource.com/324043
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>