UPSTREAM: mesh: Fix PMKID to match the standard
IEEE Std 802.11-2012 11.3.5.4 specifies the PMKID for SAE-derived keys
as:

   L((commit-scalar + peer-commit-scalar) mod r, 0, 128)

This is already calculated in the SAE code when the PMK is derived, but
not saved anywhere. Later, when generating the PMKID for plink action
frames, the definition for PMKID from 11.6.1.3 is incorrectly used.
Correct this by saving the PMKID when the key is generated and use it
subsequently.

[cherry picked from 6c33eed3ee7fd6bd9c561295e001a6b63adbb88d]
[git remote add hostap git://w1.fi/srv/git/hostap.git]

BUG=b:26058138
TEST=../platform/ap-daemons/autotest/tools/run_lab_test.py --board=whirlwind suite:jetstream_sanity
	test_that --label=pool:wificell suite:wifi_matfunc

Change-Id: Ieea711ee4026c33be1e8a410f1461e4e94ca4a97
Signed-off-by: Bob Copeland <me@bobcopeland.com>
Reviewed-on: https://chromium-review.googlesource.com/326492
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>