UPSTREAM: mesh: Do not clear link state on driver event if exchange was started
If the local driver event for a new peer candidate arrived only after
the peer had already initiated the peering exchange, we used to clear
the link state. This resulted in the already completed (or in progress)
exchange getting abandoned and a new exchange initiated. This is not
desirable since the already started (or even completed) exchange can be
used. Clear the link state only when adding the new STA entry for the
first time, i.e., use the same !sta->my_lid condition in handling the
driver event similarly to how the peer initiated cases were already
handled.

[cherry-pick from cc64fe7b9e3104c4cb990eb410930676de749e02]
[git remote add hostap git://w1.fi/srv/git/hostap.git]

BUG=b:26058138
TEST=../platform/ap-daemons/autotest/tools/run_lab_test.py --board=whirlwind suite:jetstream_sanity
	test_that --label=pool:wificell suite:wifi_matfunc

Change-Id: Ic4dc5c442f9d25d410afc6e536d75c22a7927226
Signed-off-by: Jouni Malinen <j@w1.fi>
Reviewed-on: https://chromium-review.googlesource.com/326497
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>