UPSTREAM: mesh: Do not force another peering exchange on driver event
If the local driver indicated a peer candidate event when the peer had
already initiated peering exchange in open mesh case, we used to force a
new exchange to be started instead of allowing the previously started
exchange to complete. This is not desirable, so make this initiation of
the new exchange conditional on there not being an already started (or
successfully completed) exchange.

[cherry pick from 27446e471c5e08913e8230b90ffc6c2a5659d10a]
[git remote add hostap git://w1.fi/srv/git/hostap.git]

BUG=b:26058138
TEST=../platform/ap-daemons/autotest/tools/run_lab_test.py --board=whirlwind suite:jetstream_sanity
	test_that --label=pool:wificell suite:wifi_matfunc

Change-Id: I6e5fec16d6c79552c764ff2af77ba61dbde4eb4c
Signed-off-by: Jouni Malinen <j@w1.fi>
Reviewed-on: https://chromium-review.googlesource.com/326498
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>