UPSTREAM: AP: Fix Deauth/Disassoc TX status timeout handling
The ap_sta_deauth_cb and ap_sta_disassoc_cb eloop timeouts are used to
clear a disconnecting STA from the kernel driver if the STA did not ACK
the Deauthentication/Disassociation frame from the AP within two
seconds. However, it was possible for a STA to not ACK such a frame,
e.g., when the disconnection happened due to hostapd pruning old
associations from other BSSes and the STA was not on the old channel
anymore. If that same STA then started a new authentication/association
with the BSS, the two second timeout could trigger during this new
association and result in the STA entry getting removed from the kernel.

Fix this by canceling these eloop timeouts when receiving an indication
of a new authentication or association.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
(cherry picked from commit 9e8fde21348bbdcd8c6d9892820106f93d7738f1)

BUG=b:26058138
TEST=../platform/ap-daemons/autotest/tools/run_lab_test.py --board=whirlwind suite:jetstream_sanity
	test_that --label=pool:wificell suite:wifi_matfunc

Change-Id: Ibf3347f897fb7abd2bc4dcc45563cced10ac4d7c
Reviewed-on: https://chromium-review.googlesource.com/326191
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: John Walker <walker@google.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>